+++ libs/cardano-protocol-tpraos/test/Test/Cardano/Protocol/Binary/CddlSpec.hs
@@ -8,27 +8,27 @@
 module Test.Cardano.Protocol.Binary.CddlSpec (spec) where
 
 import Cardano.Ledger.Allegra (AllegraEra)
+import Cardano.Ledger.Allegra.HuddleSpec (allegraCDDL)
 import Cardano.Ledger.Alonzo (AlonzoEra)
+import Cardano.Ledger.Alonzo.HuddleSpec (alonzoCDDL)
 import Cardano.Ledger.Binary.Group (CBORGroup)
 import Cardano.Ledger.Core
 import Cardano.Ledger.Mary (MaryEra)
+import Cardano.Ledger.Mary.HuddleSpec (maryCDDL)
 import Cardano.Ledger.Shelley (ShelleyEra)
+import Cardano.Ledger.Shelley.HuddleSpec (shelleyCDDL)
 import Cardano.Protocol.Crypto (StandardCrypto)
 import Cardano.Protocol.TPraos.BHeader (BHBody, BHeader)
 import Cardano.Protocol.TPraos.OCert (OCert)
 import Codec.CBOR.Cuddle.Huddle (Huddle)
 import qualified Data.ByteString.Lazy as BSL
 import Test.Cardano.Ledger.Allegra.Binary.Cddl (readAllegraCddlFiles)
-import Test.Cardano.Ledger.Allegra.CDDL (allegraCDDL)
 import Test.Cardano.Ledger.Alonzo.Binary.Cddl (readAlonzoCddlFiles)
-import Test.Cardano.Ledger.Alonzo.CDDL (alonzoCDDL)
 import Test.Cardano.Ledger.Binary.Cddl (beforeAllCddlFile, cddlRoundTripCborSpec)
 import Test.Cardano.Ledger.Binary.Cuddle
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Mary.Binary.Cddl (readMaryCddlFiles)
-import Test.Cardano.Ledger.Mary.CDDL (maryCDDL)
 import Test.Cardano.Ledger.Shelley.Binary.Cddl (readShelleyCddlFiles)
-import Test.Cardano.Ledger.Shelley.CDDL (shelleyCDDL)
 import Test.Cardano.Protocol.Binary.Annotator ()
 import Test.Cardano.Protocol.Binary.Cddl (
   cddlBlockSpec,