+++ eras/dijkstra/impl/CHANGELOG.md
@@ -62,6 +62,7 @@
 
 ### `testlib`
 
+* Remove `huddle-cddl` and the `CDDL` modules.
 * Re-export `Test.Cardano.Ledger.Conway.Binary.Golden`
 * Remove CDDL `certificate` redefinition to reuse from conway.
 * Add CDDL exports for `plutus_v4_script`, `dijkstra_native_script`, `script_require_guard`